% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_bdd.R
\name{bdd_to_config}
\alias{bdd_to_config}
\title{bdd_to_config}
\usage{
bdd_to_config(site_url, bdd = file.path(find.package("SEO"), "mabase.sqlite"))
}
\arguments{
\item{site_url}{url website to find}

\item{bdd}{the sqlite database path}
}
\description{
get  url config from bdd
}

