% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{compare_position}
\alias{compare_position}
\title{compare_position}
\usage{
compare_position(site_url, to = Sys.time(), from = to - hours(24),
  bdd = file.path(find.package("SEO"), "mabase.sqlite"), shiny = TRUE)
}
\arguments{
\item{site_url}{url website to find}

\item{to}{start time}

\item{from}{end time}

\item{bdd}{the sqlite database path}

\item{shiny}{booleen are we in a shiny app}
}
\description{
evolution between 2 dates
}

