% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_bdd.R
\name{config_to_bdd}
\alias{config_to_bdd}
\title{config_to_bdd}
\usage{
config_to_bdd(site_url, max = 100, timer = 10, validite = 24 * 60 * 60,
  bdd = file.path(find.package("SEO"), "mabase.sqlite"),
  source = "https://www.google.com")
}
\arguments{
\item{site_url}{url website to find}

\item{max}{maximum position to scan}

\item{timer}{time in second between each query}

\item{validite}{cache validity}

\item{bdd}{the sqlite database path}

\item{source}{google's url}
}
\description{
save url config to bdd
}

