% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_run.R
\name{daily_run}
\alias{daily_run}
\title{daily_run}
\usage{
daily_run(site_url, word_list, max = 100, timer = 10, validite = 24 * 60 *
  60, bdd = file.path(find.package("SEO"), "mabase.sqlite"),
  source = "https://www.google.com")
}
\arguments{
\item{site_url}{url website to find}

\item{word_list}{query to search in googl}

\item{max}{maximum position to scan}

\item{timer}{time in second between each query}

\item{validite}{cache validity}

\item{bdd}{the sqlite database path}

\item{source}{google's url}
}
\description{
find the daily position for a url for a list of keyword
}

