% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{escargot}
\alias{escargot}
\title{escargot}
\usage{
escargot(start, max)
}
\arguments{
\item{start}{google page number to start with}

\item{max}{maximum position to scan}
}
\description{
generate optiml page order search
}

