% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{extract_serp}
\alias{extract_serp}
\title{extract_serp}
\usage{
extract_serp(word, start = 0, source = "https://www.google.com",
  timer = 10, validite = 86400, cache_only = FALSE)
}
\arguments{
\item{word}{word to track}

\item{start}{which google page do we have to parse}

\item{source}{google's URL}

\item{timer}{time in second between each query}

\item{validite}{cache validity}

\item{cache_only}{boolean if TRUE we dont ask google}
}

