% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generate_serp}
\alias{generate_serp}
\title{generate_serp}
\usage{
generate_serp(from = as.Date("2015-05-05"), to = Sys.Date(),
  six_mots = c("mot 1", "autre mot", "mot 2", "tjrs un mot", "mot 1",
  "encore un mot"))
}
\arguments{
\item{from}{date depart}

\item{to}{date fin}

\item{six_mots}{vector of words}
}
\description{
genre serp classiques
}

