% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_all_position}
\alias{get_all_position}
\title{get_all_position}
\usage{
get_all_position(site_url, word_list, bdd = file.path(find.package("SEO"),
  "mabase.sqlite"))
}
\arguments{
\item{site_url}{url website}

\item{word_list}{words to track}

\item{bdd}{the sqlite database path}
}
\description{
retrieve all knowned position
}

