% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_best_position}
\alias{get_best_position}
\title{get_best_position}
\usage{
get_best_position(site_url, word, bdd = file.path(find.package("SEO"),
  "mabase.sqlite"))
}
\arguments{
\item{site_url}{url website to find}

\item{word}{query to search in google}

\item{bdd}{the sqlite database path}
}
\description{
retrieve best known position
}

