% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_concurence}
\alias{get_concurence}
\title{get_concurence}
\usage{
get_concurence(site_url, word, max = 2, history = 1, timer = 10,
  validite = Inf, cache_only = TRUE, cible = FALSE)
}
\arguments{
\item{site_url}{url website}

\item{word}{word to track}

\item{max}{maximum position}

\item{history}{start point}

\item{timer}{timer}

\item{validite}{cache validity}

\item{cache_only}{booleen if TRUE we dont ask google}

\item{cible}{booleen if TRUE search just around history}
}
\description{
retrieve all knowned position
}

