% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_date_position}
\alias{get_date_position}
\title{get_date_position}
\usage{
get_date_position(site_url, ladate = Sys.time(),
  bdd = file.path(find.package("SEO"), "mabase.sqlite"))
}
\arguments{
\item{site_url}{url website to find}

\item{ladate}{the date to retrieve information}

\item{bdd}{the sqlite database path}
}
\description{
retrieve known position at date
}

