% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{get_position}
\alias{get_position}
\title{get_position}
\usage{
get_position(site_url, word, max = 29, history = 0, timer = 10,
  validite = 24 * 60 * 60, source = "https://www.google.com")
}
\arguments{
\item{site_url}{url website to find}

\item{word}{query to search in googl}

\item{max}{maximum position to scan}

\item{history}{last known position in google}

\item{timer}{time in second between each query}

\item{validite}{cache validity}

\item{source}{google's url}
}
\description{
find the position url for a keyword
}

