% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{list_url_base}
\alias{list_url_base}
\title{list_url_base}
\usage{
list_url_base(bdd = file.path(find.package("SEO"), "mabase.sqlite"))
}
\arguments{
\item{bdd}{the sqlite database path}
}
\description{
list all url in base
}

