% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{plot_evolution_mot}
\alias{plot_evolution_mot}
\title{plot_evolution_mot}
\usage{
plot_evolution_mot(site_url, word_list, bdd = file.path(find.package("SEO"),
  "mabase.sqlite"))
}
\arguments{
\item{site_url}{the site url}

\item{word_list}{the list of words}

\item{bdd}{the sqlite database path}
}
\description{
draw evolution of an url position for one word
}

