% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{url_decode}
\alias{url_decode}
\title{url_decode}
\usage{
url_decode(url, to = c("<", ">", "#", "\%", "{", "}", "|", "\\\\", "^", "~",
  "[", "]", "`", ";", "/", "?", ":", "@", "=", "&", "$", "+", "\\"", " "),
  from = c("\%3C", "\%3E", "\%23", "\%25", "\%7B", "\%7D", "\%7C", "\%5C",
  "\%5E", "\%7E", "\%5B", "\%5D", "\%60", "\%3B", "\%2F", "\%3F", "\%3A",
  "\%40", "\%3D", "\%26", "\%24", "\%2B", "\%22", "\%20"))
}
\arguments{
\item{url}{the url}

\item{to}{output}

\item{from}{input}
}
\description{
decode url
}

