% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_bdd.R
\name{word_list_to_bdd}
\alias{word_list_to_bdd}
\title{word_list_to_bdd}
\usage{
word_list_to_bdd(site_url, word_list, bdd = file.path(find.package("SEO"),
  "mabase.sqlite"), append = FALSE)
}
\arguments{
\item{site_url}{url website to find}

\item{word_list}{query to search in googl}

\item{bdd}{the sqlite database path}

\item{append}{do we have to add word or to replace word_list}
}
\description{
save url word_list to bdd
}

